requirements.txt
anthropic>=0.18.0
ollama>=0.1.0
pymupdf>=1.23.0
python-dotenv>=1.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
streamlit>=1.31.0
